incorrect parsing of recursive directoriers + timestamp for failsafe copy

+++ basic_cms/management/commands/compress_cms_images.py
@@ -1,11 +1,13 @@
 import tempfile
 import os
+import time
+import datetime
+from optparse import make_option
 
 from django.core.management.base import BaseCommand  # , CommandError
 from django.core.files.storage import default_storage
 from basic_cms import settings
 from image_diet import squeeze
-from optparse import make_option
 
 
 class Command(BaseCommand):
@@ -22,32 +24,39 @@ class Command(BaseCommand):
     )
 
     def handle(self, *args, **options):
+        timestamp = time.time()
+        timestamp = datetime.datetime.fromtimestamp(timestamp).strftime('-%Y-%m-%d-%H:%M:%S')
 
         def compress_files(data, dirtree):
             for f in data[1]:  # files from listdir
-                path = os.sep.join(dirtree)
-                path = os.path.join(path, f)
-                try:
-                    path = default_storage.path(path)
-                    squeeze(path)
-                except NotImplementedError:
-                    if path[-1:] != os.sep:
-                        pf = default_storage.open(path, 'rwb')
-                        print("Processing %s" % pf.name)
-                        image = pf.read()
-                        tmpfilehandle, tmpfilepath = tempfile.mkstemp()
-                        tmpfilehandle = os.fdopen(tmpfilehandle, 'wb')
-                        tmpfilehandle.write(image)
-                        tmpfilehandle.close()
-                        squeeze(tmpfilepath)
-                        tmpfilehandle = open(tmpfilepath)
-                        pf.close()
-                        default_storage.save(path, tmpfilehandle)
-                        os.remove(tmpfilepath)
+                if f and f.lower().endswith(('.png', '.jpg', '.jpeg', '.gif')):  # sometimes it is [u'']
+                    path = os.sep.join(dirtree)
+                    path = os.path.join(path, f)
+                    # failsafe copy of file
+                    copy = default_storage.open(path, 'rb')
+                    default_storage.save(path + timestamp, copy)
+                    copy.close()
+                    try:
+                        path = default_storage.path(path)
+                        squeeze(path)
+                    except NotImplementedError:
+                        if path[-1:] != os.sep:
+                            pf = default_storage.open(path, 'rwb')
+                            print("Processing %s" % pf.name)
+                            image = pf.read()
+                            tmpfilehandle, tmpfilepath = tempfile.mkstemp()
+                            tmpfilehandle = os.fdopen(tmpfilehandle, 'wb')
+                            tmpfilehandle.write(image)
+                            tmpfilehandle.close()
+                            squeeze(tmpfilepath)
+                            tmpfilehandle = open(tmpfilepath)
+                            pf.close()
+                            default_storage.save(path, tmpfilehandle)
+                            os.remove(tmpfilepath)
 
             for d in data[0]:  # directories from list_dir
                 dirtree.append(d)
-                d = default_storage.listdir(d)
+                d = default_storage.listdir(os.sep.join(dirtree))
                 compress_files(d, dirtree)
                 dirtree.pop()  # remove last item, not needed anymore
